requirements.txt
streamlit
deep-translator
python-Levenshtein